+++ ticket-service/src/main/java/com/epam/training/ticketservice/commandprompt/commands/RoomCommand.java
@@ -3,6 +3,8 @@ package com.epam.training.ticketservice.commandprompt.commands;
 import com.epam.training.ticketservice.commandprompt.commands.checks.CheckService;
 import com.epam.training.ticketservice.core.room.RoomService;
 import com.epam.training.ticketservice.core.room.model.RoomDto;
+import com.epam.training.ticketservice.core.user.UserService;
+import com.epam.training.ticketservice.core.user.persistence.entity.User;
 import lombok.AllArgsConstructor;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
@@ -15,15 +17,19 @@ import java.util.stream.Collectors;
 public class RoomCommand {
     private final RoomService roomService;
     private final CheckService checkService;
+    private final UserService userService;
 
     @ShellMethod(key = "create room",value = "Creating a room")
     public String createRoom(String name, Integer seatRows, Integer seatColumns) throws Exception {
         if (!roomService.getRooms().isEmpty() && checkService.isExistingRoom(name)) {
             throw new Exception("There already is a room with that name");
         }
-        RoomDto roomDto = new RoomDto(name, seatRows, seatColumns);
-        roomService.createRoom(roomDto);
-        return roomDto.toString();
+        if (userService.describe().isPresent() && userService.describe().get().getRole().equals(User.Role.ADMIN)) {
+            RoomDto roomDto = new RoomDto(name, seatRows, seatColumns);
+            roomService.createRoom(roomDto);
+            return roomDto.toString();
+        }
+        return null;
     }
 
     @ShellMethod(key = "update room",value = "Updating room")
@@ -31,8 +37,11 @@ public class RoomCommand {
         if (!checkService.isExistingRoom(name)) {
             return "There is no room with that name";
         }
-        roomService.updateRoom(name, seatRows, seatColumns);
-        return new RoomDto(name, seatRows, seatColumns) + " updated";
+        if (userService.describe().isPresent() && userService.describe().get().getRole().equals(User.Role.ADMIN)) {
+            roomService.updateRoom(name, seatRows, seatColumns);
+            return new RoomDto(name, seatRows, seatColumns) + " updated";
+        }
+        return null;
     }
 
     @ShellMethod(key = "delete room",value = "Deleting room")
@@ -40,8 +49,11 @@ public class RoomCommand {
         if (!checkService.isExistingRoom(name)) {
             return "There is no room with that name";
         }
-        roomService.deleteRoom(name);
-        return "Room " + name + " successfully deleted";
+        if (userService.describe().isPresent() && userService.describe().get().getRole().equals(User.Role.ADMIN)) {
+            roomService.deleteRoom(name);
+            return "Room " + name + " successfully deleted";
+        }
+        return null;
     }
 
     @ShellMethod(key = "list rooms",value = "List of rooms")
